More algorithms; merge to OSV soon
This build includes a lot of experimental new image tools, a new algorithm variant, and another algorithm that is entirely new, which works by getting clusters of color. Those clusters need to be analyzed in a way that would require a lot of time and may not look all that good. I might do that later. For now, this setup has been tested with every single texture included with OSV where stone is the background and not a single one of them has any issues. Backgrounds with more complex colors (such as netherrack, which has two distinct colors throughout and not so much just different shades of the same color) may not always work as well. That being said, this will allow me to remove almost every single texture from the mod (that I haven't already taken out). 

Which is pretty sick. 

Am I gonna give up on Conquest textures? Yeah, maybe. Not sure. Might ask the devs for permission to use their textures again. See if they change their minds.